requirements.txt
moviepy==1.0.3
gTTS==2.3.2
requests==2.31.0
beautifulsoup4==4.13.0
schedule==1.1.0
python-dotenv==1.0.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.102.0
transformers==4.35.0
